requirements.txt
Flask
requests
openai-whisper
